requirements.txt
# Text Mining Tool Dependencies

# Core dependencies
nltk>=3.8.1
flask>=2.3.0
flask-wtf>=1.2.0
requests>=2.31.0

# File handling
python-docx>=1.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.0
networkx>=3.1.0

# Web interface
dash>=2.13.0
dash-bootstrap-components>=1.4.2

# Utilities
tqdm>=4.65.0
joblib>=1.3.1

# Testing
pytest>=7.4.0

# Advanced NLP (optional)
# spacy>=3.6.0
# scikit-learn>=1.3.0
# gensim>=4.3.0

# Performance Optimization
joblib==1.3.2
tqdm==4.66.1 